Refactor JS code

+++ static/js/endgame/game.js
@@ -1,4 +1,6 @@
-class Game {
+import State from "./state.js";
+
+export default class Game {
     constructor(fen, dtm) {
         this.fen = fen;
         this.chess = new Chess(fen);
@@ -15,7 +17,7 @@ class Game {
             promotion: "q",
         });
 
-        if (move != null) {
+        if (move !== null && move !== undefined) {
             if (!this.isLastMove()) {
                 this.truncate();
             }
@@ -60,7 +62,7 @@ class Game {
             if (this.chess.in_draw()) {
                 return "½-½";
             } else if (this.chess.in_checkmate()) {
-                return this.chess.turn() == "w" ? "0-1" : "1-0";
+                return this.chess.turn() === "w" ? "0-1" : "1-0";
             }
         }
     }
@@ -74,7 +76,7 @@ class Game {
     }
 
     isLastMove() {
-        return this.currentMove == this.states.length - 1;
+        return this.currentMove === this.states.length - 1;
     }
 
     getDTZ() {
